Move title function to BaseItemUtils
